+++ shared/src/index.ts
@@ -2,5 +2,5 @@ export * from './models/common';
 export * from './models/room';
 export * from './models/uno';
 export * from './games';
-export * from './events/game-events';
+export * from './games/uno/events';
 export * from './protocol/messages';